requirements.txt
numpy
pandas
scikit-learn
sentence-transformers
rank-bm25
pdfminer.six
transformers>=4.39.0
accelerate
peft
torch --extra-index-url https://download.pytorch.org/whl/cpu
streamlit
